Rename quam components
