+++ Quantum-Control-Applications/Superconducting/quam/Two-Flux-Tunable-Transmons/quam_components/__init__.py
@@ -1,11 +1,11 @@
-from .Transmon_component import *
-from .ReadoutResonator_component import *
-from .FluxLine_component import *
-from .quam_components import *
+from .transmon import *
+from .readout_resonator import *
+from .flux_line import *
+from .quam import *
 
 __all__ = [
-    *Transmon_component.__all__,
-    *ReadoutResonator_component.__all__,
-    *FluxLine_component.__all__,
-    *quam_components.__all__,
+    *transmon.__all__,
+    *readout_resonator.__all__,
+    *flux_line.__all__,
+    *quam.__all__,
 ]
